examples/speech_to_text/requirements.txt
pillow>=4.3.0
ipython[all]
tqdm
sox
ruamel.yaml
jupyterlab
tqdm
boto3
requests
six
ipdb
h5py
html2text
nltk
progressbar
matplotlib
wget 
tensorboardX
pandas
onnx
wget
num2words
librosa
inflect
kaldi-io
marshmallow
unidecode
sentencepiece
boto3
matplotlib
h5py
youtokentome
docrep
torch-stft
